# frontend/src/lib/backend/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "Sur la route", "version": "0.1.0" },
	"paths": {
		"/api/login/access-token": {
			"post": {
				"tags": ["login"],
				"summary": "Access Token",
				"description": "OAuth2 compatible token login, get an access token for future requests",
				"operationId": "login-access_token",
				"requestBody": {
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": { "$ref": "#/components/schemas/Body_login-access_token" }
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/Token" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/login/test-token": {
			"post": {
				"tags": ["login"],
				"summary": "Test Token",
				"description": "Test access token",
				"operationId": "login-test_token",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
						}
					}
				},
				"security": [{ "OAuth2PasswordBearer": [] }]
			}
		},
		"/api/password-recovery/{email}": {
			"post": {
				"tags": ["login"],
				"summary": "Recover Password",
				"description": "Password Recovery",
				"operationId": "login-recover_password",
				"parameters": [
					{
						"name": "email",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "title": "Email" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/reset-password/": {
			"post": {
				"tags": ["login"],
				"summary": "Reset Password",
				"description": "Reset password",
				"operationId": "login-reset_password",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/NewPassword" } }
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/password-recovery-html-content/{email}": {
			"post": {
				"tags": ["login"],
				"summary": "Recover Password Html Content",
				"description": "HTML Content for Password Recovery",
				"operationId": "login-recover_password_html_content",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"parameters": [
					{
						"name": "email",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "title": "Email" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": { "text/html": { "schema": { "type": "string" } } }
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
...
		"/api/directory/organisations/": {
			"post": {
				"tags": ["organisations"],
				"summary": "Create Organisation",
				"description": "Create an organisation.",
				"operationId": "organisations-create_organisation",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/OrganisationCreate" } }
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/OrganisationPublic" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"get": {
				"tags": ["organisations"],
				"summary": "Read Organisations",
				"description": "Read paginated organisations.",
				"operationId": "organisations-read_organisations",
				"parameters": [
					{
						"name": "q",
						"in": "query",
						"required": false,
						"schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Q" }
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"maximum": 100,
							"exclusiveMinimum": 0,
							"default": 100,
							"title": "Limit"
						}
					},
					{
						"name": "offset",
						"in": "query",
						"required": false,
						"schema": { "type": "integer", "minimum": 0, "default": 0, "title": "Offset" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PagedResponse_OrganisationPublic_" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/directory/organisations/{id}": {
			"get": {
				"tags": ["organisations"],
				"summary": "Read Organisation By Id",
				"description": "Read an organisation by its id.",
				"operationId": "organisations-read_organisation_by_id",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Id" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/OrganisationPublic" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"patch": {
				"tags": ["organisations"],
				"summary": "Update Organisation",
				"description": "Update an organisation.",
				"operationId": "organisations-update_organisation",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Id" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/OrganisationUpdate" } }
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/UpdateResponse_OrganisationPublic_" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["organisations"],
				"summary": "Delete Organisation",
				"description": "Delete an organisation.",
				"operationId": "organisations-delete_organisation",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Id" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/DeleteResponse" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
...
			"Body_login-access_token": {
				"properties": {
					"grant_type": {
						"anyOf": [{ "type": "string", "pattern": "password" }, { "type": "null" }],
						"title": "Grant Type"
					},
					"username": { "type": "string", "title": "Username" },
					"password": { "type": "string", "title": "Password" },
					"scope": { "type": "string", "title": "Scope", "default": "" },
					"client_id": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Client Id"
					},
					"client_secret": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Client Secret"
					}
				},
				"type": "object",
				"required": ["username", "password"],
				"title": "Body_login-access_token"
			},
			"DeleteResponse": {
				"properties": {
					"success": { "type": "boolean", "title": "Success" },
					"message": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Message" }
				},
				"type": "object",
				"required": ["success"],
				"title": "DeleteResponse",
				"description": "Response schema for any paged API."
			},
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": { "$ref": "#/components/schemas/ValidationError" },
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
			"Message": {
				"properties": { "message": { "type": "string", "title": "Message" } },
				"type": "object",
				"required": ["message"],
				"title": "Message"
			},
			"NewPassword": {
				"properties": {
					"token": { "type": "string", "title": "Token" },
					"new_password": {
						"type": "string",
						"maxLength": 40,
						"minLength": 8,
						"title": "New Password"
					}
				},
				"type": "object",
				"required": ["token", "new_password"],
				"title": "NewPassword"
			},
			"OrganisationCreate": {
				"properties": { "name": { "type": "string", "title": "Name" } },
				"type": "object",
				"required": ["name"],
				"title": "OrganisationCreate"
			},
			"OrganisationPublic": {
				"properties": {
					"id": { "type": "string", "format": "uuid", "title": "Id" },
					"name": { "type": "string", "title": "Name" }
				},
				"type": "object",
				"required": ["id", "name"],
				"title": "OrganisationPublic"
			},
			"OrganisationUpdate": {
				"properties": {
					"name": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Name" }
				},
				"type": "object",
				"title": "OrganisationUpdate"
			},
			"PagedResponse_OrganisationPublic_": {
				"properties": {
					"total": { "type": "integer", "title": "Total" },
					"limit": { "type": "integer", "title": "Limit" },
					"offset": { "type": "integer", "title": "Offset" },
					"results": {
						"items": { "$ref": "#/components/schemas/OrganisationPublic" },
						"type": "array",
						"title": "Results"
					}
				},
				"type": "object",
				"required": ["total", "limit", "offset", "results"],
				"title": "PagedResponse[OrganisationPublic]"
			},
...
			"Token": {
				"properties": {
					"access_token": { "type": "string", "title": "Access Token" },
					"token_type": { "type": "string", "title": "Token Type", "default": "bearer" }
				},
				"type": "object",
				"required": ["access_token"],
				"title": "Token"
			},
...
			"UpdateResponse_OrganisationPublic_": {
				"properties": {
					"success": { "type": "boolean", "title": "Success" },
					"data": {
						"anyOf": [{ "$ref": "#/components/schemas/OrganisationPublic" }, { "type": "null" }]
					},
					"message": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Message" }
				},
				"type": "object",
				"required": ["success"],
				"title": "UpdateResponse[OrganisationPublic]"
			},
...
			"UserPublic": {
				"properties": {
					"email": { "type": "string", "maxLength": 255, "format": "email", "title": "Email" },
					"is_active": { "type": "boolean", "title": "Is Active", "default": true },
					"is_superuser": { "type": "boolean", "title": "Is Superuser", "default": false },
					"full_name": {
						"anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
						"title": "Full Name"
					},
					"id": { "type": "string", "format": "uuid", "title": "Id" }
				},
				"type": "object",
				"required": ["email", "id"],
				"title": "UserPublic"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
						"type": "array",
						"title": "Location"
					},
					"msg": { "type": "string", "title": "Message" },
					"type": { "type": "string", "title": "Error Type" }
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			}
		},
		"securitySchemes": {
			"OAuth2PasswordBearer": {
				"type": "oauth2",
				"flows": { "password": { "scopes": {}, "tokenUrl": "/api/login/access-token" } }
			}
		}
	}
}